+++ src/main/java/com/authorizationapi/authorizationapi/service/estructura/ParticipanteGrupoService.java
@@ -1,12 +1,18 @@
 package com.authorizationapi.authorizationapi.service.estructura;
 
 
+import com.authorizationapi.authorizationapi.crosscutting.utils.UtilUUID;
+import com.authorizationapi.authorizationapi.crosscutting.utils.exception.AuthorizationServiceException;
+import com.authorizationapi.authorizationapi.crosscutting.utils.messages.UtilMessagesService;
 import com.authorizationapi.authorizationapi.domain.estructura.ParticipanteGrupo;
 import com.authorizationapi.authorizationapi.repository.ParticipanteGrupoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
 @Service
 public final class ParticipanteGrupoService {
 
@@ -14,6 +20,14 @@ public final class ParticipanteGrupoService {
     private ParticipanteGrupoRepository repository;
 
     public void asignarGrupo(ParticipanteGrupo participante) {
+        UUID identificador;
+        Optional<ParticipanteGrupo> participanteOptional;
+        do {
+            identificador = UtilUUID.generateNewUUID();
+            participanteOptional = repository.findById(identificador);
+        } while (participanteOptional.isPresent());
+        participante.setIdentificador(identificador);
+
         repository.save(participante);
     }
 
@@ -23,7 +37,13 @@ public final class ParticipanteGrupoService {
 
 
     public void eliminar(ParticipanteGrupo participante) {
-        repository.delete(participante);
+        Optional<ParticipanteGrupo> participanteOptional = repository.findById(participante.getIdentificador());
+
+        if (participanteOptional.isPresent()) {
+            repository.delete(participante);
+        } else {
+            throw AuthorizationServiceException.create(UtilMessagesService.ServiceParticipanteGrupo.PARTICIPANTE_GRUPO_NO_ENCONTRADO);
+        }
     }
 
 }
