+++ src/main/java/com/authorizationapi/authorizationapi/service/organizacion/AdministradorOrganizacionEncargadoService.java
@@ -1,5 +1,9 @@
 package com.authorizationapi.authorizationapi.service.organizacion;
 
+import com.authorizationapi.authorizationapi.crosscutting.utils.UtilBoolean;
+import com.authorizationapi.authorizationapi.crosscutting.utils.UtilUUID;
+import com.authorizationapi.authorizationapi.crosscutting.utils.exception.AuthorizationServiceException;
+import com.authorizationapi.authorizationapi.crosscutting.utils.messages.UtilMessagesService;
 import com.authorizationapi.authorizationapi.domain.organizacion.AdministradorOrganizacionEncargado;
 import com.authorizationapi.authorizationapi.repository.AdministradorOrganizacionEncargadoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,20 +18,28 @@ public final class AdministradorOrganizacionEncargadoService {
     @Autowired
     private AdministradorOrganizacionEncargadoRepository repository;
     public void concederPermisos(AdministradorOrganizacionEncargado administrador) {
+        UUID identificador;
+        Optional<AdministradorOrganizacionEncargado> administradorOptional;
+        do {
+            identificador = UtilUUID.generateNewUUID();
+            administradorOptional = repository.findById(identificador);
+        } while (administradorOptional.isPresent());
+        administrador.setIdentificador(identificador);
+
         repository.save(administrador);
     }
 
-    public void cambiarEstado(UUID identificador, AdministradorOrganizacionEncargado nuevoEstadoAdministrador) {
+    public void cambiarEstado(UUID identificador) {
         Optional<AdministradorOrganizacionEncargado> administradorOptional = repository.findById(identificador);
 
         if (administradorOptional.isPresent()) {
             AdministradorOrganizacionEncargado administradorExistente = administradorOptional.get();
-            administradorExistente.setActivo(nuevoEstadoAdministrador.isActivo());
+            administradorExistente.setActivo(UtilBoolean.getOpposite(administradorExistente.isActivo()));
 
             repository.save(administradorExistente);
         } else {
 
-            throw new RuntimeException("Administrador organizacion no encontrado con el identificador: " + identificador);
+            throw AuthorizationServiceException.create(UtilMessagesService.ServiceAdministradorOrganizacionEncargado.ADMINISTRADOR_NO_ENCONTRADO_IDENTIFICADOR + identificador);
         }
     }
 
@@ -36,7 +48,13 @@ public final class AdministradorOrganizacionEncargadoService {
     }
 
     public void eliminar(AdministradorOrganizacionEncargado administrador) {
-        repository.delete(administrador);
+        Optional<AdministradorOrganizacionEncargado> administradorOptional = repository.findById(administrador.getIdentificador());
+
+        if (administradorOptional.isPresent()) {
+            repository.delete(administrador);
+        } else {
+            throw AuthorizationServiceException.create(UtilMessagesService.ServiceAdministradorOrganizacionEncargado.ADMINISTRADOR_NO_ENCONTRADO);
+        }
     }
 
 }
